requirements.txt
# cspell:ignore gunicorn psycopg whitenoise
Django==5.2.4
Pillow==11.3.0
gunicorn==23.0.0
waitress==3.0.2
psycopg2-binary==2.9.10
dj-database-url==3.0.1
whitenoise==6.9.0
python-decouple==3.8
